feat: use proper postgres error
