+++ go/services/multigateway/handler/handler_test.go
@@ -219,7 +219,7 @@ func TestPreparedStatementHandling(t *testing.T) {
 	// 6. Describe fails after close
 	_, err = handler.HandleDescribe(ctx, conn, 'S', "stmt1")
 	require.Error(t, err)
-	require.Contains(t, err.Error(), "does not exist")
+	require.True(t, mterrors.IsErrorCode(err, mterrors.PgSSInvalidSQLStatementName))
 
 	// 7. Empty query fails
 	err = handler.HandleParse(ctx, conn, "empty", "", nil)
@@ -252,7 +252,7 @@ func TestPortalHandling(t *testing.T) {
 	// 1. Bind fails for non-existent statement
 	err = handler.HandleBind(ctx, conn, "portal1", "nonexistent", nil, nil, nil)
 	require.Error(t, err)
-	require.Contains(t, err.Error(), "does not exist")
+	require.True(t, mterrors.IsErrorCode(err, mterrors.PgSSInvalidSQLStatementName))
 
 	// 2. Bind stores portal in connection state
 	params := [][]byte{[]byte("42")}
@@ -278,7 +278,7 @@ func TestPortalHandling(t *testing.T) {
 		return nil
 	})
 	require.Error(t, err)
-	require.Contains(t, err.Error(), "does not exist")
+	require.True(t, mterrors.IsErrorCode(err, mterrors.PgSSInvalidCursorName))
 
 	// 6. Close removes portal from connection state
 	err = handler.HandleClose(ctx, conn, 'P', "portal1")
@@ -287,7 +287,7 @@ func TestPortalHandling(t *testing.T) {
 	// 7. Describe fails after close
 	_, err = handler.HandleDescribe(ctx, conn, 'P', "portal1")
 	require.Error(t, err)
-	require.Contains(t, err.Error(), "does not exist")
+	require.True(t, mterrors.IsErrorCode(err, mterrors.PgSSInvalidCursorName))
 }
 
 // TestPreparedStatementConsolidation tests that same queries share the same
